web/pyproject.toml
[project]
name = "drf-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "django-cors-headers>=4.9.0",
    "dotenv>=0.9.9",
    "django-filter>=25.2",
    "psycopg2-binary>=2.9.11",
    "pillow>=11.3.0",
    "drf-yasg>=1.21.11",
    "gunicorn>=23.0.0",
    "django-jazzmin>=3.0.1",
    "ruff>=0.14.1",
    "pytz>=2025.2",
]
